partially switched to generate and return
